Create html file
Created index.html file and populated it with the structure of the
roaming-terra landing page project. All sections of the landing page;
the header, the various subsections of the content section, and the
footer, should be defined in such a way to position them using flexbox.
Potential edits may be performed down the line if positioning is
rendered difficult. Additionally, edits will be made down the line
to personalize this landing page for the roaming-terra brand.