Create css document
Css document was linked to html document, and elements were styled and
positioned for the header and section-1 of the content. The remaining
three section of the content and the footer still need to be worked on
in terms of the css. I also realized I don't like the way text collapses
when I shrink the window, so I wrote some questions as notes in the css
document to ask them in the discord server of the Odin Project at a
later time.