Stylize and position elements in section 2 content
Section 2 of the content section of the roaming-terra landing page has
been stylized and elements have been positioned. Images have yet to be
placed within the borders of the cards. Additionally, index file has
been edited. button 2 was changed to button 1, which was originally
a spelling error in the html. div containers have also been added to
section 2 subsection 2 in order to better position elements. Notes are
still present in the style.css document, meaning questions still need
to be asked in the discord server of the Odin Project.