Stylize and position elements of section 3 content
Section 3 of the content section has been stylized and elements have
been positioned. I added a note to the bottom of the style.css document
to describe the fact that I need to decide how to collapse elements when
the window shrinks. This relates somewhat to the notes written at the
beginning of the style.css document. This can be worked on after all
sections have been stylized with css.