Stylize and position element section 4 content
Section 4 of the content section has been stylized and the elements
have been positioned. I made the sign-up button of this section's card
wrap to the bottom when the screen shrinks. Additionally, I added a
couple of div containers in the index.html for section 4
for better positioning, I changed a div element under the
call-to-action-container to a header element for being able to reference
it in the style.css file, and I changed the width of both buttons on the
landing page to a minimum width of 90px, because
I don't want the button to morph in size as the
window changes size. I like the idea of buttons not changing size and
wrapping below when the window shrinks, so I will stick to this. I also
think collapsing some text is ok. I'll have some logs to look over as
I get to the editing of the css for shrinking window purposes. Also
keep in mind that I am using the odin project landing page as
inspiration for how to move elements as the window shrinks.