Modify all sections to consider shrinking window
All sections were modified to keep a clear appearance while the window
shrinks vertically. The header and section 1 css was modified to allow
the flex items to grow with flex grow and fill up the flex container
either equally in the case of section 1, or having only one flex item
fill up the remaining space in the case of the header. Flex wrap was
used as well to keep a clear appearance. Padding and flex wrap were
added as well to section 2 to assure a clear appearance. Div elements
were added in various sections for specificity and containing purposes.
Section 4 was already styled in an appropriate way to consider the
horizontal shrinking of the window. For section 3 however, the right
and left padding were changed to a percentage so that when the window
is completely horizontally minimized, the quotee is directly under the
quote and inline. This may change when the quote changes. For this
reason, Section 3's css may eventually be modified in a similar fashion
to section 1.