Organize css
Removed redundant comments, as well as the comment at the beginning
regarding shrinking text when window shrinks, because I learned that
it is better to wrap elements than to shrink elements. Css was reviewed
and made more efficient and organized to better help legibility for
other developers. Regarding the HTML, I added a class .container to
subsections 1 and 2 in section 1 to make css more efficient. I also
reviewed the colors and edited some of them to better reflect the
desired final product.