Fix section 4 of the content
Fixed the padding of section 4 to make it in line with the header and
section 1. The blue padding of the card now shrinks with the shrinking
window, until the window reaches its smallest size, resulting in the
blue card taking up the entire horizontal space of said small window.
This was done for stylistic purposes. A class .button-container was
created to have nested in it the .button-2 class in the html for flex
purposes. This was done to make the button element and text element
within the card use the flex property and grow to take up a given space.