Personalize landing page
Html and css files were modified, and a folder of royalty free images
was created to personalize the landing page. Additionally, a favicon was
embedded in the html, the header was made sticky to keep it at the top
of the window while scrolling, css rules were defined to allow for the
image in section 1 to shrink as thew window shrinks, and the footer was
edited to give credit to myself as the creator.